.gitmodules: Add ignore=dirty for libbpf
We run 'make install' inside the libbpf subdirectory which will mark it as
'dirty' as far as the parent repo is concerned. Since we're not doing any
libbpf development in this repository, just ignoring it should not bring
any issues; we'll still get a notification if the committed content of the
submodule changes.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>