pping: Refactor cleanup code
Add cleanup code to load_attach_bpfprogs function, so it should always
unpin tc program, and additionally detach the tc and xdp programs in
case of any failure.

Also rename some of the functions for attaching/detaching tc programs
to be more consistent with the xdp ones.

Still need to keep a copy of most of the cleanup code in main as well,
as the tc and xdp programs also need to be detached on program
shutdown or if later functions fail.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>