pping: Genarlize test script slightly
Make which machines the test should be set up on configurable. Also
change naming/path conventions to refer to the machines as M1 - M3
instead of VM1 - VM3 as the machines are not necessarily
virtual.

Also reduce threshold for when the CPU is detected as under load. In
general it would probably be better to try and match up the timestamps
to find the correct part of the CPU-utalization trace, although that
would require handling some additional complexities (like the
wrap-around at midnight).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>