AF_XDP-interaction: rename --poll-mode to --wakeup-mode
It is confusing to name the option what will wait for packets
on the file descriptor, for --poll, just because the function
call have this name.  It confusing as AF_XDP users often want
to busy-poll for packets (for max performance reasons).

Name the new option --wakeup instead of --wait as the effect
of waiting for the FD is that the kernel needs to wakeup
the userspace process.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>