pping: Add additional monitoring on test machines
Run sar instead of mpstat to monitor a much wider range of
attributes (not just CPU) on the machines invovled in the test. This
can for example be used to get some network metrics independent from
the traffic generator (currently iperf3). After compression, the sar
files are still relatively small (compared to the output
from (e)PPing).

For now, use sadf to also create a JSON file with CPU metrics in a
format very similar to the one produced by mpstat in order to keep it
compatible with the plotting scripts. In future, should probably let
plotting scripts themselves run sadf to retrieve the relevant data
from the sar file.

Also made slight changes to mpstat_viz.py to allow it to work with
both normal mpstat JSON files as well the JSON files produced by the
sadf -j command (which are very similar).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>