pping: Improve program cleanup on shutdown
Explicitly stop the thread which performs periodical map cleanup on
shutdown, before attempting to free up any resources it might use.

For the cleanup order to make more sense, also setup the perf-buffer
before setting up the periodical map cleaning, so that the thread can
be stopped as the first part of the cleanup. This also matches better
with the next couple of commits where map cleaning debug information
will be pushed through the perf-buffer.

Finally, move the addition of the signalhandler earlier in the code to
eliminate a window where it was possible to terminate the program
without relevant cleanup code having a chance to run.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>