global_func_test: Add initial test program
A fairly simple program (parses incoming IMCP packets and print out
their sequence number with printk) that is mainly intended to be used
as a basis for figuring out how to get global functions
working.

The --log-level (-l) flag can be used to set the log-level of the
verifier. For any log-level > 0, the verifier log will be printed even
if the program passes the verifier.

The existence of a couple of defines determines how the BPF code is
structured.

USE_GLOBAL_FUNC:
If defined, the parse_packet function will be treated as a global
function which will be verified independently. Right now the code will
not get past the verifier in this mode, and the goal is to get that
working.

USE_COMMON_CONTEXT:
If defined, the parse_packet function will take a pointer to a struct
common_context instead of an xdp_md. The idea is that a common_context
could easily be created for ex. a tc BPF program as well (which use
__sk_buff as their context), making it possible to use parse_packet
from either an XDP or tc program (similar to how struct
parsing_context is used in pping).

While the verifier will reject the code regardless of
USE_COMMON_CONTEXT when using global functions, it does so for
different reasons. When trying to use the common context, it will fail
verifying the parse_packet function entirely as it does not recognize
data and data_end as packet pointers, and therefore checking against
data_end is not sufficient to allow directly accessing packet
memory. When directly using the xdp_md (which has type PTR_TO_CTX) it
does not have this issue and the parse_packet function is successfully
verified as a global function, but it then fails calling the function
from the XDP program (it complains that the type of the first argument
is PTR_TO_CTX while the function expects a PTR_TO_STACK, which
confuses me as it just sucessfully verified the function assuming the
first argument was PTR_TO_CTX).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>