pkt-loop-filter: Add fallback if we can't pin bpf_link for kprobe
Support for bpf_link-based attaching of kprobes was added to kernel 5.15
with commit: b89fbfbb854c ("bpf: Implement minimal BPF perf link"). Prior
to this, it is not possible to pin kprobe attachments in bpffs, which
causes the pkt-loop-filter to fail. Add a fallback where we just keep
running in the foreground to keep the probe alive if bpf_link pinning
fails.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>