pkt-loop-filter: Keep running in background instead of foreground
When pinning of the bpf_link fails, we keep running to keep the PID alive.
However, staying in the foreground causes problems with scripts that
expects the setup to finish running; so fork into the background instead
and write a PID file so we can kill the running instance on unload.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>