pkt-loop-filter: Fix gratuitous ARP handling
The exception for gratuitous ARPs are only supposed to be for entries that
would otherwise be dropped due to the loop filtering logic. In addition, we
should record egress gratuitous ARPs and make sure they don't trigger the
exception when looping back (this is 'rule 4' of the openvswitch SLB
bonding logic).

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>