Issue with the metadata section
I am not sure if I am doing something wrong or if XDQ is not storing the
adjusted metadata pointer after the bpf_xdp_adjust_meta function call.

When I try to access the data_meta pointer in the dequeue hook, I get a
different metadata pointer address from the XDP hook. I expect it to
point eight bytes back from the data pointer. However, it points to the
same address as the data pointer.

Signed-off-by: Frey Alfredsson <freysteinn@freysteinn.com>